embedding-requirements.txt
torch>=2.0.0
sentence-transformers>=2.2.0
Pillow>=9.0.0
numpy>=1.21.0
